Bug Fix, built game (did not send to git hub)
Music added to main menu
Bug fix on built game, cursor was visable added Cursor.visible = false on our starting script, then also added it to the pause menu. but iwth a true and false value.
Moved Files.
I believe this is at the point where we can submit it.